# claroz-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for the Claroz social platform",
        "title": "Claroz API",
        "termsOfService": "https://claroz.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://claroz.com/support",
            "email": "lucas.little@claroz.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://claroz.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "schemes": ["http"],
    "paths": {
...
        "/users/me": {
            "get": {
                "security": [{"Bearer": []}],
                "description": "Get the currently authenticated user's details",
                "produces": ["application/json"],
                "tags": ["users"],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Retrieve a user by their UUID",
                "produces": ["application/json"],
                "tags": ["users"],
                "summary": "Get user by ID",
                "parameters": [{
                    "type": "string",
                    "description": "User ID (UUID)",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [{"Bearer": []}],
                "description": "Update an existing user's information",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["users"],
                "summary": "Update user details",
                "parameters": [{
                    "type": "string",
                    "description": "User ID (UUID)",
                    "name": "id",
                    "in": "path",
                    "required": true
                }, {
                    "description": "Updated user details",
                    "name": "user",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/models.User"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [{"Bearer": []}],
                "description": "Delete a user by their ID",
                "produces": ["application/json"],
                "tags": ["users"],
                "summary": "Delete user",
                "parameters": [{
                    "type": "string",
                    "description": "User ID (UUID)",
                    "name": "id",
                    "in": "path",
                    "required": true
                }],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "username": {
                    "type": "string",
                    "example": "lucas"
                },
                "email": {
                    "type": "string",
                    "example": "lucas.little@claroz.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Lucas Little"
                },
                "bio": {
                    "type": "string",
                    "example": "Software engineer at Claroz"
                },
                "avatar": {
                    "type": "string",
                    "example": "https://claroz.com/avatars/lucas.jpg"
                },
                "did": {
                    "type": "string",
                    "example": "did:web:claroz.com:lucas"
                },
                "handle": {
                    "type": "string",
                    "example": "@lucas"
                },
                "federation_type": {
                    "type": "string",
                    "example": "local"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-26T00:35:27Z"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-26T00:35:27Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Type \"Bearer\" followed by a space and JWT token."
        }
    }
}